RUNNING
• Function:
   - added prefs as properties on Function objects (call corresponding object.prefs.pref setters and getters)

• Process:
  - fixed bug involving import of LearningSignal

 • State:
   - fixed bug in parsing lambda function when used as specification of param value

[FROM Minstrel]
